Hydrator-1316 get log for preview
