+++ cdap-app-fabric/src/main/java/co/cask/cdap/app/preview/PreviewRunner.java
@@ -17,8 +17,9 @@
 package co.cask.cdap.app.preview;
 
 import co.cask.cdap.api.metrics.MetricTimeSeries;
+import co.cask.cdap.internal.app.store.RunRecordMeta;
+import co.cask.cdap.proto.id.ProgramRunId;
 import com.google.gson.JsonElement;
-import org.apache.twill.api.logging.LogEntry;
 
 import java.util.List;
 import java.util.Map;
@@ -69,8 +70,14 @@ public interface PreviewRunner {
   List<MetricTimeSeries> getMetrics();
 
   /**
-   * Get the logs for the preview run represented by this {@link PreviewRunner}.
-   * @return the logs
+   * Get the run id of the program executed as a part of preview.
+   * @return the {@link ProgramRunId} associated with the preview
    */
-  List<LogEntry> getLogs();
+  ProgramRunId getProgramRunId();
+
+  /**
+   * Get the run record of the program executed as a part of preview.
+   * @return the {@link RunRecordMeta} associated with the preview
+   */
+  RunRecordMeta getRunRecord();
 }
